+++ CHANGELOG.md
@@ -72,6 +72,7 @@ These changes are available on the `master` branch, but have not yet been releas
 - Added `suppress` and `allowed_mentions` parameters to `Webhook` and
   `InteractionResponse` edit methods.
   ([#2138](https://github.com/Pycord-Development/pycord/pull/2138))
+- Added `wait_end` coroutine method to `VoiceClient`.
 
 ### Changed